POC/LLAMA MCP Agent/requirements.txt
langchain
langchain-core
langchain-community 
langchain-ollama
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
ollama
